+++ packages/fcl/src/wallet-utils/on-fcl-message.js
@@ -1,0 +1,11 @@
+export const onFclMessage = (msg, cb = () => {}) => {
+  const internal = e => {
+    if (typeof e.data !== "object") return
+    if (e.data.type !== msg) return
+    delete e.data.body.interaction
+
+    cb(e.data)
+  }
+  window.addEventListener("message", internal)
+  return () => window.removeEventListener("message", internal)
+}